+++ cli/docs/changelog.md
@@ -1,8 +1,19 @@
 # TBD
+### Features
+* The enclave ID argument to `enclave inspect` is now tab-completable
+* Added a "Debugging User Issues" section to the README
+* Added a wrapper over the `CobraCmd` objects that we're using to create CLI commands, so that we have a centralized place to add autocompletion
+
 ### Changes
 * Upgraded the engine server to 1.9.1
 * Upgraded the kurtosis core to 1.37.1
 * Upgraded the metrics library to 0.1.2
+* Switched the `enclave inspect` to the new command framework
+* Cleaned up our Kurtosis-custom logic wrapping the Cobra commands
+
+### Fixes
+* Fix the Kurtosis completion
+* The CLI version mismatch warning messages are printed to STDERR so as not to interfere with any other command output (e.g. `completion`)
 
 # 0.9.0
 ### Features